Fix compile warnings on 32 bit systems

+++ sym/tests/aes_cbc_dec_tester.c
@@ -110,8 +110,9 @@ static int test_decrypt_cbc(const struct lc_sym *aes, const char *name)
 	int ret;
 	LC_SYM_CTX_ON_STACK(aes_cbc, aes);
 
-	printf("AES CBC ctx %s (%s implementation) len %lu\n", name,
-	       aes == lc_aes_cbc_c ? "C" : "accelerated", LC_SYM_CTX_SIZE(aes));
+	printf("AES CBC ctx %s (%s implementation) len %u\n", name,
+	       aes == lc_aes_cbc_c ? "C" : "accelerated",
+	       (unsigned int)LC_SYM_CTX_SIZE(aes));
 
 	unpoison(key256, sizeof(key256));
 	memcpy(in, in256, sizeof(in256));
