+++ sym/tests/aes_ctr_tester.c
@@ -121,8 +121,9 @@ static int test_xcrypt_ctr(const struct lc_sym *aes, const char *name,
 	int ret;
 	LC_SYM_CTX_ON_STACK(aes_ctr, aes);
 
-	printf("AES CTR ctx %s (%s implementation) len %lu\n", name,
-	       aes == lc_aes_ctr_c ? "C" : "accelerated", LC_SYM_CTX_SIZE(aes));
+	printf("AES CTR ctx %s (%s implementation) len %u\n", name,
+	       aes == lc_aes_ctr_c ? "C" : "accelerated",
+	       (unsigned int)LC_SYM_CTX_SIZE(aes));
 
 	memcpy(in2, in256, sizeof(in256));
 	ret = test_xcrypt_ctr_one(xcrypt, aes_ctr, key256, sizeof(key256), in2);
