+++ sym/tests/aes_kw_tester.c
@@ -110,8 +110,9 @@ static int test_kw(const struct lc_sym *aes, const char *name)
 	int ret;
 	LC_SYM_CTX_ON_STACK(aes_kw, aes);
 
-	printf("AES KW ctx %s (%s implementation) len %lu\n", name,
-	       aes == lc_aes_kw_c ? "C" : "accelerated", LC_SYM_CTX_SIZE(aes));
+	printf("AES KW ctx %s (%s implementation) len %u\n", name,
+	       aes == lc_aes_kw_c ? "C" : "accelerated",
+	       (unsigned int)LC_SYM_CTX_SIZE(aes));
 
 	ret = test_encrypt_kw_one(aes_kw, key128, sizeof(key128), pt128,
 				  sizeof(pt128), ct128, iv128);
